==> dataset_building/debug_notebooks/3.split_train_test.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
import pickle

# Load Dataset

data = pd.read_csv('../../data/data.csv')
data = data.drop(columns=['state_fips_code', 'county_fips_code', 'case_month'])
data.pop_estimate_2019 = data.pop_estimate_2019.str.replace(',', '').astype(float)
data = data.astype(float)
data

# Convert Dataset

Z = data.iloc[:, :].values.astype(float)
del data

X, y = Z[:, 2:], Z[:, :2]
del Z

# Split Dataset

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=20210927)

# Write Datasets

with open('../../data/data_train.pt', 'wb') as file:
    pickle.dump((X_train, y_train), file)

with open('../../data/data_test.pt', 'wb') as file:
    pickle.dump((X_test, y_test), file)

# Test Functionality

with open('../../data/data_train.pt', 'rb') as file:
    X_train_loaded, y_train_loaded = pickle.load(file)

X_train_loaded.shape, y_train.shape

with open('../../data/data_test.pt', 'rb') as file:
    X_test_loaded, y_test_loaded = pickle.load(file)

X_test_loaded.shape, y_test.shape
